Added Kruskal.c
Kruskal's algorithm is a greedy algorithm for constructing a Minimum Spanning Tree (MST). Unlike Prim's algorithm, which starts with a single node and grows the tree gradually, Kruskal's algorithm takes a different approach.

Initialization:

Begin with an empty set of edges.
Sort all the edges in the graph by their weights in non-decreasing order.
Adding Edges:

Starting with the edge of the smallest weight, add edges to the MST one by one.
Ensure that adding an edge does not create a cycle in the MST. If adding an edge creates a cycle, skip it.
Termination:

Continue adding edges until the MST contains (V-1) edges, where V is the number of nodes in the graph.